requirements.txt
Flask==2.0.3
requests==2.27.1
python-dotenv